+++ example/lib/i18n/locale/en_us.dart
@@ -118,4 +118,5 @@ Map<String, String> enUS = {
   '索引栏': 'Index Bar',
   '业务组件': 'Business Components',
   '我是自定义内容': 'I am custom content',
+  '折叠面板':'Collapse'
 };