+++ example/lib/router/routes.dart
@@ -1,5 +1,6 @@
 import 'package:example/pages/better_button_page.dart';
 import 'package:example/pages/better_cell_page.dart';
+import 'package:example/pages/better_collapse_page.dart';
 import 'package:example/pages/better_date_picker_page.dart';
 import 'package:example/pages/better_index_bar_page.dart';
 import 'package:example/pages/better_marquee_page.dart';
@@ -30,4 +31,5 @@ final routes = [
   GetPage(name: "/betterMarquee", page: () => const BetterMarqueePage()),
   GetPage(name: "/betterSwiper", page: () => const BetterSwiperPage()),
   GetPage(name: "/betterIndexBar", page: () => const BetterIndexBarPage()),
+  GetPage(name: "/betterCollapse", page: () => const BetterCollapsePage()),
 ];